+++ src/main/java/com/lucent/skeleton/repository/ScriptRepository.java
@@ -4,7 +4,7 @@ package com.lucent.skeleton.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.lucent.skeleton.entities.GlobalShopifyScript;
 
-public interface ScriptRepository extends CrudRepository<GlobalShopifyScript, Integer> {
+public interface ScriptRepository extends CrudRepository<GlobalShopifyScript, Long> {
     @Override
     Iterable<GlobalShopifyScript> findAll();
