+++ src/main/java/com/lucent/skeleton/repository/StorePlanRepository.java
@@ -6,6 +6,6 @@ import com.lucent.skeleton.entities.StorePlanDetails;
 
 import java.util.List;
 
-public interface StorePlanRepository extends CrudRepository<StorePlanDetails, Integer> {
-    List<StorePlanDetails> findByStoreId(Integer storeId);
+public interface StorePlanRepository extends CrudRepository<StorePlanDetails, Long> {
+    List<StorePlanDetails> findByStoreId(Long storeId);
 }
